Move tethering runtime out of app-fabric
