+++ cdap-runtime-ext-tethering/.gitignore
@@ -1,0 +1,1 @@
+data/
