+++ pom.xml
@@ -2898,6 +2898,7 @@
         <module>cdap-proto</module>
         <module>cdap-runtime-ext-dataproc</module>
         <module>cdap-runtime-ext-emr</module>
+        <module>cdap-runtime-ext-tethering</module>
         <module>cdap-runtime-ext-remote-hadoop</module>
         <module>cdap-runtime-spi</module>
         <module>cdap-securestore-ext-cloudkms</module>
